update readme and fix bug in cloudinary

+++ src/utils/cloudinary.js
@@ -9,20 +9,25 @@ cloudinary.config({
 
 const uploadOnCloudinary = async (localFilePath) => {
     try {
-        if (!localFilePath) return null
-        // upload the file on cloudinary
+        if (!localFilePath) return null;
+
         const response = await cloudinary.uploader.upload(localFilePath, {
-           resource_type: "auto" 
-        })
-        //file has been uploaded succesfully
-        // console.log("File is uploaded on cloudinary", response.url);
-        fs.unlinkSync(localFilePath)
+            resource_type: "auto"
+        });
+
+        if (fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return response;
-        
+
     } catch (error) {
-        fs.unlinkSync(localFilePath) // remove the locally saved temporary file as the upload operation got failed
+        if (localFilePath && fs.existsSync(localFilePath)) {
+            fs.unlinkSync(localFilePath);
+        }
+
         return null;
-    } 
-}
+    }
+};
 
 export { uploadOnCloudinary, cloudinary }
